day 3 bug fix
